requirements.txt
pytesseract
langchain
langchain-ollama
pillow
mss
pynput
PyQt5
